+++ cpp/core/memory/MemoryAllocator.h
@@ -33,10 +33,10 @@ class MemoryAllocator {
 
   virtual bool allocate(int64_t size, void** out) = 0;
   virtual bool allocateZeroFilled(int64_t nmemb, int64_t size, void** out) = 0;
-  virtual bool allocateAligned(uint16_t alignment, int64_t size, void** out) = 0;
+  virtual bool allocateAligned(uint64_t alignment, int64_t size, void** out) = 0;
 
   virtual bool reallocate(void* p, int64_t size, int64_t newSize, void** out) = 0;
-  virtual bool reallocateAligned(void* p, uint16_t alignment, int64_t size, int64_t newSize, void** out) = 0;
+  virtual bool reallocateAligned(void* p, uint64_t alignment, int64_t size, int64_t newSize, void** out) = 0;
 
   virtual bool free(void* p, int64_t size) = 0;
 
@@ -67,11 +67,11 @@ class ListenableMemoryAllocator final : public MemoryAllocator {
 
   bool allocateZeroFilled(int64_t nmemb, int64_t size, void** out) override;
 
-  bool allocateAligned(uint16_t alignment, int64_t size, void** out) override;
+  bool allocateAligned(uint64_t alignment, int64_t size, void** out) override;
 
   bool reallocate(void* p, int64_t size, int64_t newSize, void** out) override;
 
-  bool reallocateAligned(void* p, uint16_t alignment, int64_t size, int64_t newSize, void** out) override;
+  bool reallocateAligned(void* p, uint64_t alignment, int64_t size, int64_t newSize, void** out) override;
 
   bool free(void* p, int64_t size) override;
 
@@ -93,11 +93,11 @@ class StdMemoryAllocator final : public MemoryAllocator {
 
   bool allocateZeroFilled(int64_t nmemb, int64_t size, void** out) override;
 
-  bool allocateAligned(uint16_t alignment, int64_t size, void** out) override;
+  bool allocateAligned(uint64_t alignment, int64_t size, void** out) override;
 
   bool reallocate(void* p, int64_t size, int64_t newSize, void** out) override;
 
-  bool reallocateAligned(void* p, uint16_t alignment, int64_t size, int64_t newSize, void** out) override;
+  bool reallocateAligned(void* p, uint64_t alignment, int64_t size, int64_t newSize, void** out) override;
 
   bool free(void* p, int64_t size) override;
